Refactoring: moved parseCsvList to Strings and added more test cases.
