+++ fhir-server/src/main/java/au/csiro/pathling/fhirpath/TerminologyUtils.java
@@ -1,13 +1,7 @@
 package au.csiro.pathling.fhirpath;
 
-import static java.util.function.Predicate.not;
-
 import au.csiro.pathling.fhirpath.element.ElementPath;
 import au.csiro.pathling.fhirpath.literal.CodingLiteralPath;
-import java.util.List;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import org.hl7.fhir.r4.model.Enumerations.FHIRDefinedType;
 
@@ -15,23 +9,6 @@ import org.hl7.fhir.r4.model.Enumerations.FHIRDefinedType;
  * Terminology helper functions
  */
 public interface TerminologyUtils {
-
-  /**
-   * Leniently parses a list of coma separated values. Trims the code strings and filters out empty
-   * values.
-   *
-   * @param csvList a coma separated list of equivalence codes
-   * @param converter a function that converts single value string to the desired type T
-   * @param <T> the type of elements to produce.
-   * @return the list of converted values of type T.
-   */
-  @Nonnull
-  static <T> List<T> parseCsvList(@Nonnull final String csvList, final @Nonnull
-      Function<String, T> converter) {
-    return Stream.of(csvList.split(",")).map(String::trim).filter(not(String::isEmpty))
-        .map(converter).collect(
-            Collectors.toList());
-  }
 
   /**
    * Checks if a path if a codeable concept element.
