+++ fhir-server/src/main/java/au/csiro/pathling/utilities/Strings.java
@@ -6,9 +6,15 @@
 
 package au.csiro.pathling.utilities;
 
+import static java.util.function.Predicate.not;
+
+import java.util.List;
 import java.util.Random;
+import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 
 /**
@@ -56,4 +62,20 @@ public abstract class Strings {
     return matcher.matches();
   }
 
+  /**
+   * Leniently parses a list of coma separated values. Trims the strings and filters out empty
+   * values.
+   *
+   * @param csvList a coma separated list of equivalence codes
+   * @param converter a function that converts single value string to the desired type T
+   * @param <T> the type of elements to produce.
+   * @return the list of converted values of type T.
+   */
+  @Nonnull
+  public static <T> List<T> parseCsvList(@Nonnull final String csvList, final @Nonnull
+      Function<String, T> converter) {
+    return Stream.of(csvList.split(",")).map(String::trim).filter(not(String::isEmpty))
+        .map(converter).collect(
+            Collectors.toList());
+  }
 }
